feat: refactor custom views to their packages
