+++ app/src/main/res/layout/activity_finish.xml
@@ -84,7 +84,7 @@
             </androidx.cardview.widget.CardView>
 
 
-            <aeropresscipe.divinelink.aeropress.customviews.RecipeCardView
+            <aeropresscipe.divinelink.aeropress.components.recipecard.RecipeCardView
                 android:id="@+id/card"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
