+++ app/src/main/res/layout/view_swipe_recipe_card.xml
@@ -19,7 +19,7 @@
         app:sal_isFullActionSupported="false"
         app:sal_shouldVibrateOnQuickAction="false">
 
-        <aeropresscipe.divinelink.aeropress.customviews.RecipeCardView
+        <aeropresscipe.divinelink.aeropress.components.recipecard.RecipeCardView
             android:id="@+id/card"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
